add dynamic generating sensor-channels...
add  particularly device initial state
bug fixes and more

+++ extensions/binding/org.eclipse.smarthome.binding.digitalstrom/src/main/java/org/eclipse/smarthome/binding/digitalstrom/internal/lib/structure/devices/Device.java
@@ -448,7 +448,7 @@ public interface Device extends GeneralDeviceInformations {
      * @param sceneNumber
      * @param prio
      */
-    public void checkSceneConfig(Short sceneNumber, int prio);
+    public void checkSceneConfig(Short sceneNumber, short prio);
 
     /**
      * Sets the given output mode as new output mode of this {@link Device}.
